fix(abi): propagate errno on membrane Deny paths and add fallback allocation tracking
ABI shims previously returned null/error silently when the membrane
policy denied a call, leaving errno unset. Callers that check errno
(e.g., getpwnam, getgrnam, fopen) would see stale values and
misattribute failures.

Add set_abi_errno() helpers to grp, pwd, process, resolv, pthread,
stdio, time, and wchar ABI modules that set errno = EACCES on every
Deny path, giving callers a clear signal that the operation was
policy-blocked rather than a system error.

In malloc_abi, add a lock-free open-addressed hash table
(FALLBACK_ALLOC_TABLE_SLOTS = 4096 AtomicUsize slots) to track which
allocations went through native libc (bootstrap/reentrant paths) vs
the membrane arena. Uses multiplicative hashing, CAS-based insertion
with tombstone reuse, and CAS-based removal. This allows free/realloc
to route deallocations to the correct allocator, preventing
use-after-free and double-free when mixing allocation sources.

+++ crates/frankenlibc-abi/src/wchar_abi.rs
@@ -1031,7 +1031,8 @@ pub unsafe extern "C" fn wmemcmp(s1: *const u32, s2: *const u32, n: usize) -> c_
         let mut res = 0;
         for i in 0..cmp_len {
             if a[i] != b[i] {
-                res = if a[i] < b[i] { -1 } else { 1 };
+                // wchar_t is signed (i32) on Linux; compare as signed values.
+                res = if (a[i] as i32) < (b[i] as i32) { -1 } else { 1 };
                 break;
             }
         }
